test(live): merge the parameter-only Stop and mobile-toolbar specs
Eight specs, two behaviors.

The four `composer-stop-*` files differed only in the single session/update
the fake emitted before an identical 30s hold, and were byte-identical from
the Stop click onward. They failed together for any break in the cancel
path, and each paid its own `aoe serve` boot, agent subprocess and browser
context to vary one field. One spec now drives all four openings as four
scripted turns against one session: 4 x ~9s becomes ~11s.

The four `mobile-toolbar-*` files differed only in a (button name, expected
bytes) pair on one code path, four full serve boots to assert four
constants. Merged into a table in one spec. `MobileTerminalToolbar.test.tsx`
and the mocked `tests/mobile-toolbar.spec.ts` cover the buttons' presence
but assert nothing about what they emit, so the mapping keeps live coverage.

Coverage matrix collapsed to one entry per merged spec.
