--- web/tests/live/acp-stories/mobile-toolbar-escape.spec.ts
@@ -1,57 +1,0 @@
-// User story: tapping the Esc button on the mobile terminal toolbar
-// sends "\x1b" to the PTY.
-
-import { test as base, expect, devices } from "@playwright/test";
-import { spawnAoeServe, listSessions, seedSessionViaAoeAdd } from "../../helpers/aoeServe";
-
-base.use({ ...devices["iPhone 13"] });
-
-base("mobile toolbar Esc sends ESC", async ({ page }, testInfo) => {
-  const serve = await spawnAoeServe({
-    authMode: "none",
-    workerIndex: testInfo.workerIndex,
-    parallelIndex: testInfo.parallelIndex,
-    seedFn: seedSessionViaAoeAdd({ title: "story-mobile-esc" }),
-  });
-
-  try {
-    const sessions = await listSessions(serve.baseUrl);
-    const seeded = sessions.find((s) => s.title === "story-mobile-esc");
-    if (!seeded) throw new Error("seeded session 'story-mobile-esc' missing");
-    const sessionId = seeded.id;
-
-    await page.addInitScript(() => {
-      const w = window as unknown as { __WS_SENT__: string[] };
-      w.__WS_SENT__ = [];
-      const origSend = WebSocket.prototype.send;
-      WebSocket.prototype.send = function (data: BufferSource | string) {
-        try {
-          if (data instanceof ArrayBuffer) {
-            w.__WS_SENT__.push(new TextDecoder().decode(new Uint8Array(data)));
-          } else if (ArrayBuffer.isView(data)) {
-            w.__WS_SENT__.push(new TextDecoder().decode(data as unknown as Uint8Array));
-          } else if (typeof data === "string") {
-            w.__WS_SENT__.push(data);
-          }
-        } catch {
-          // swallow
-        }
-        return origSend.call(this, data as never);
-      };
-    });
-
-    await page.goto(`${serve.baseUrl}/session/${encodeURIComponent(sessionId)}`);
-
-    const esc = page.getByRole("button", { name: "Escape" });
-    await expect(esc).toBeVisible({ timeout: 15_000 });
-    await esc.click();
-
-    await expect
-      .poll(async () => await page.evaluate(() => (window as unknown as { __WS_SENT__: string[] }).__WS_SENT__), {
-        timeout: 5_000,
-      })
-      .toContain("\x1b");
-  } finally {
-    await serve.stop();
-  }
-});
